kobject: move internal functions to own header
Move all internal kobject APIs to own header and do not expose them with
the rest of the public API.
